Change build structure
