+++ .vscode/launch.json
@@ -11,8 +11,8 @@
       "skipFiles": ["<node_internals>/**"],
       "cwd": "${workspaceFolder}",
       "console": "externalTerminal",
-      "program": "${workspaceFolder}/src/.build/bidiServer/index.js",
-      "outFiles": ["${workspaceFolder}/src/.build/bidiServer/**/*.js"],
+      "program": "${workspaceFolder}/lib/cjs/bidiServer/bidiServer/index.js",
+      "outFiles": ["${workspaceFolder}/lib/cjs/bidiServer/bidiServer/**/*.js"],
       "args": ["--headless=false"],
       "env": {
         "DEBUG": "*",
