+++ package.json
@@ -4,23 +4,22 @@
   "description": "An implementation of the WebDriver BiDi protocol for Chromium implemented as a JavaScript layer translating between BiDi and CDP, running inside a Chrome tab.",
   "main": "index.ts",
   "scripts": {
-    "build": "npm run prettier && npm run clean && npm run build-mapper && npm run build-server",
-    "build-mapper": "rollup --config src/bidiMapper/rollup.config.js",
-    "build-server": "tsc -b src/tsconfig.json",
-    "clean": "rimraf src/.build",
+    "build": "npm run clean && tsc -b src/tsconfig.json && npm run rollup",
+    "rollup": "rollup --config configs/rollup.config.js",
+    "clean": "tsc -b --clean src/tsconfig.json && rimraf lib/iife",
     "e2e": "npm run e2e-headless",
     "e2e-headful": "npm run server-no-build -- --headless=false & npm run e2e-only",
     "e2e-headless": "npm run server-no-build & npm run e2e-only",
     "e2e-only": "python3 -m pytest --rootdir=tests",
     "lint": "npx prettier --check .",
     "server": "npm run build && npm run server-no-build --",
-    "server-no-build": "node src/.build/bidiServer/index.js",
+    "server-no-build": "node lib/cjs/bidiServer/bidiServer/index.js",
     "test": "npm run unit",
     "prettier": "npx prettier --write .",
     "unit": "cross-env TS_NODE_PROJECT='src/tsconfig.json' mocha"
   },
   "files": [
-    "src/.build/"
+    "lib"
   ],
   "repository": {
     "type": "git",
@@ -41,9 +40,7 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-typescript": "^8.3.2",
     "@types/argparse": "^2.0.10",
     "@types/chai": "^4.3.1",
     "@types/chai-as-promised": "^7.1.5",
@@ -60,7 +57,6 @@
     "prettier": "2.6.2",
     "rimraf": "^3.0.2",
     "rollup": "^2.74.1",
-    "rollup-plugin-string": "^3.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "sinon": "^14.0.0",
     "terser": "^5.14.2",
