--- src/bidiMapper/scripts.d.ts
@@ -1,4 +1,0 @@
-declare module '*.es' {
-  const content: string;
-  export default content;
-}
