+++ src/bidiServer/bidiServerRunner.ts
@@ -17,7 +17,7 @@
 
 import websocket from 'websocket';
 import http from 'http';
-import {ITransport} from './utils/transport';
+import {ITransport} from '../utils/transport';
 
 import debug from 'debug';
